@@ -2,14 +2,14 @@
 using Furball.Engine.Engine.Localization;
 using Furball.Game.Screens;
 
-namespace Furball.Game {
-    public class FurballTestGame : FurballGame {
-        public FurballTestGame() : base(new ScreenSelector()) {
-            //this.Window.AllowUserResizing = true;
-        }
+namespace Furball.Game; 
 
-        protected override void InitializeLocalizations() {
-            LocalizationManager.AddDefaultTranslation("cat", "Cat");
-        }
+public class FurballTestGame : FurballGame {
+    public FurballTestGame() : base(new ScreenSelector()) {
+        //this.Window.AllowUserResizing = true;
+    }
+
+    protected override void InitializeLocalizations() {
+        LocalizationManager.AddDefaultTranslation("cat", "Cat");
     }
 }
