@@ -10,147 +10,147 @@ using sowelipisona;
 using sowelipisona.Effects;
 using Color=Furball.Vixie.Backends.Shared.Color;
 
-namespace Furball.Game.Screens {
-    public class AudioEffectTest : TestScreen {
-        private AudioStream _testingStream;
+namespace Furball.Game.Screens; 
 
-        private DrawableTextBox _filenameTextBox;
+public class AudioEffectTest : TestScreen {
+    private AudioStream _testingStream;
 
-        public override void Initialize() {
-            base.Initialize();
+    private DrawableTextBox _filenameTextBox;
 
-            TextDrawable topText = new TextDrawable(new Vector2(1280f / 2f, 40), FurballGame.DEFAULT_FONT, "Audio Effects testing!!!", 48) {
-                OriginType = OriginType.Center
-            };
+    public override void Initialize() {
+        base.Initialize();
 
-            this.Manager.Add(topText);
+        TextDrawable topText = new TextDrawable(new Vector2(1280f / 2f, 40), FurballGame.DEFAULT_FONT, "Audio Effects testing!!!", 48) {
+            OriginType = OriginType.Center
+        };
 
-            DrawableTextBox filenameTextBox = new DrawableTextBox(new Vector2(50, 80), FurballGame.DEFAULT_FONT, 28, 350, "audio.mp3");
+        this.Manager.Add(topText);
 
-            this._filenameTextBox = filenameTextBox;
+        DrawableTextBox filenameTextBox = new DrawableTextBox(new Vector2(50, 80), FurballGame.DEFAULT_FONT, 28, 350, "audio.mp3");
 
-            this.Manager.Add(filenameTextBox);
+        this._filenameTextBox = filenameTextBox;
 
-            DrawableButton playButton = new DrawableButton(
-                new Vector2(50, 130),
-                FurballGame.DEFAULT_FONT,
-                28,
-                "Play Audio normally",
-                Color.White,
-                Color.Black,
-                Color.Black,
-                new Vector2(350, 40),
-                PlayButtonOnClick
-            );
+        this.Manager.Add(filenameTextBox);
 
-            this.Manager.Add(playButton);
+        DrawableButton playButton = new DrawableButton(
+        new Vector2(50, 130),
+        FurballGame.DEFAULT_FONT,
+        28,
+        "Play Audio normally",
+        Color.White,
+        Color.Black,
+        Color.Black,
+        new Vector2(350, 40),
+        PlayButtonOnClick
+        );
 
-            TextDrawable effectText = new TextDrawable(new Vector2(190, 180), FurballGame.DEFAULT_FONT, "Effects", 28);
+        this.Manager.Add(playButton);
 
-            this.Manager.Add(effectText);
+        TextDrawable effectText = new TextDrawable(new Vector2(190, 180), FurballGame.DEFAULT_FONT, "Effects", 28);
 
-            int currentY = 220;
+        this.Manager.Add(effectText);
 
-            DrawableButton lowPassPlayButton = new DrawableButton(
-            new Vector2(50, currentY),
-            FurballGame.DEFAULT_FONT,
-            28,
-            "Play Audio with Low Pass filter",
-            Color.White,
-            Color.Black,
-            Color.Black,
-            new Vector2(350, 40),
-            this.LowPassPlayOnClick
-            );
+        int currentY = 220;
 
-            this.Manager.Add(lowPassPlayButton);
+        DrawableButton lowPassPlayButton = new DrawableButton(
+        new Vector2(50, currentY),
+        FurballGame.DEFAULT_FONT,
+        28,
+        "Play Audio with Low Pass filter",
+        Color.White,
+        Color.Black,
+        Color.Black,
+        new Vector2(350, 40),
+        this.LowPassPlayOnClick
+        );
 
-            currentY += 60;
+        this.Manager.Add(lowPassPlayButton);
 
-            DrawableButton highPassPlayButton = new(
-            new Vector2(50, currentY),
-            FurballGame.DEFAULT_FONT,
-            28,
-            "Play Audio with High Pass filter",
-            Color.White,
-            Color.Black,
-            Color.Black,
-            new Vector2(350, 40),
-            this.HighPassPlayOnClick
-            );
+        currentY += 60;
 
-            this.Manager.Add(highPassPlayButton);
+        DrawableButton highPassPlayButton = new(
+        new Vector2(50, currentY),
+        FurballGame.DEFAULT_FONT,
+        28,
+        "Play Audio with High Pass filter",
+        Color.White,
+        Color.Black,
+        Color.Black,
+        new Vector2(350, 40),
+        this.HighPassPlayOnClick
+        );
 
-            currentY += 60;
+        this.Manager.Add(highPassPlayButton);
 
-            DrawableButton reverbPlayButton = new DrawableButton(
-                new Vector2(50, currentY),
-                FurballGame.DEFAULT_FONT,
-                28,
-                "Play Audio with Reverb",
-                Color.White,
-                Color.Black,
-                Color.Black,
-                new Vector2(350, 40), ReverbPlayOnClick
-            );
+        currentY += 60;
 
-            this.Manager.Add(reverbPlayButton);
+        DrawableButton reverbPlayButton = new DrawableButton(
+        new Vector2(50, currentY),
+        FurballGame.DEFAULT_FONT,
+        28,
+        "Play Audio with Reverb",
+        Color.White,
+        Color.Black,
+        Color.Black,
+        new Vector2(350, 40), ReverbPlayOnClick
+        );
+
+        this.Manager.Add(reverbPlayButton);
+    }
+
+    private void ReverbPlayOnClick(object sender, (MouseButton, Point) e) {
+        if (this._testingStream?.PlaybackState == PlaybackState.Playing) {
+            this._testingStream.Stop();
         }
 
-        private void ReverbPlayOnClick(object sender, (MouseButton, Point) e) {
-            if (this._testingStream?.PlaybackState == PlaybackState.Playing) {
-                this._testingStream.Stop();
-            }
+        this._testingStream = FurballGame.AudioEngine.CreateStream(this._filenameTextBox.Text);
 
-            this._testingStream = FurballGame.AudioEngine.CreateStream(this._filenameTextBox.Text);
+        ReverbAudioEffect effect = FurballGame.AudioEngine.CreateReverbEffect(this._testingStream);
 
-            ReverbAudioEffect effect = FurballGame.AudioEngine.CreateReverbEffect(this._testingStream);
+        effect.Apply();
 
-            effect.Apply();
+        this._testingStream.Volume = 0.4;
+        this._testingStream.Play();
+    }
 
-            this._testingStream.Volume = 0.4;
-            this._testingStream.Play();
+    private void LowPassPlayOnClick(object sender, (MouseButton, Point) e) {
+        if (this._testingStream?.PlaybackState == PlaybackState.Playing) {
+            this._testingStream.Stop();
         }
 
-        private void LowPassPlayOnClick(object sender, (MouseButton, Point) e) {
-            if (this._testingStream?.PlaybackState == PlaybackState.Playing) {
-                this._testingStream.Stop();
-            }
+        this._testingStream = FurballGame.AudioEngine.CreateStream(this._filenameTextBox.Text);
 
-            this._testingStream = FurballGame.AudioEngine.CreateStream(this._filenameTextBox.Text);
+        LowPassFilterAudioEffect effect = FurballGame.AudioEngine.CreateLowPassFilterEffect(this._testingStream);
 
-            LowPassFilterAudioEffect effect = FurballGame.AudioEngine.CreateLowPassFilterEffect(this._testingStream);
+        effect.Apply();
 
-            effect.Apply();
+        effect.FrequencyCutoff = 200;
 
-            effect.FrequencyCutoff = 200;
+        this._testingStream.Volume = 0.4;
+        this._testingStream.Play();
+    }
 
-            this._testingStream.Volume = 0.4;
-            this._testingStream.Play();
+    private void HighPassPlayOnClick(object sender, (MouseButton, Point) e) {
+        if (this._testingStream?.PlaybackState == PlaybackState.Playing)
+            this._testingStream.Stop();
+
+        this._testingStream = FurballGame.AudioEngine.CreateStream(this._filenameTextBox.Text);
+
+        HighPassFilterAudioEffect effect = FurballGame.AudioEngine.CreateHighPassFilterEffect(this._testingStream);
+
+        effect.Apply();
+
+        this._testingStream.Volume = 0.4;
+        this._testingStream.Play();
+    }
+
+    private void PlayButtonOnClick(object sender, (MouseButton, Point) e) {
+        if (this._testingStream?.PlaybackState == PlaybackState.Playing) {
+            this._testingStream.Stop();
         }
 
-        private void HighPassPlayOnClick(object sender, (MouseButton, Point) e) {
-            if (this._testingStream?.PlaybackState == PlaybackState.Playing)
-                this._testingStream.Stop();
-
-            this._testingStream = FurballGame.AudioEngine.CreateStream(this._filenameTextBox.Text);
-
-            HighPassFilterAudioEffect effect = FurballGame.AudioEngine.CreateHighPassFilterEffect(this._testingStream);
-
-            effect.Apply();
-
-            this._testingStream.Volume = 0.4;
-            this._testingStream.Play();
-        }
-
-        private void PlayButtonOnClick(object sender, (MouseButton, Point) e) {
-            if (this._testingStream?.PlaybackState == PlaybackState.Playing) {
-                this._testingStream.Stop();
-            }
-
-            this._testingStream        = FurballGame.AudioEngine.CreateStream(this._filenameTextBox.Text);
-            this._testingStream.Volume = 0.4;
-            this._testingStream.Play();
-        }
+        this._testingStream        = FurballGame.AudioEngine.CreateStream(this._filenameTextBox.Text);
+        this._testingStream.Volume = 0.4;
+        this._testingStream.Play();
     }
 }
