@@ -2,12 +2,12 @@ using Furball.Engine.Engine.Graphics.Drawables.Primitives;
 using Furball.Vixie.Backends.Shared;
 using Screen=Furball.Engine.Engine.Screen;
 
-namespace Furball.Game.Screens {
-    public class CircleDrawableTest : TestScreen {
-        public override void Initialize() {
-            base.Initialize();
+namespace Furball.Game.Screens; 
+
+public class CircleDrawableTest : TestScreen {
+    public override void Initialize() {
+        base.Initialize();
             
-            this.Manager.Add(new CirclePrimitiveDrawable(new(100), 50, Color.Red, 2));
-        }
+        this.Manager.Add(new CirclePrimitiveDrawable(new(100), 50, Color.Red, 2));
     }
 }
