@@ -6,31 +6,31 @@ using Furball.Engine;
 using Furball.Engine.Engine.Graphics.Drawables;
 using Furball.Vixie.Backends.Shared;
 
-namespace Furball.Game.Screens {
-    public class LoadingScreenTest : TestScreen {
-        public override bool RequireLoadingScreen => true;
+namespace Furball.Game.Screens; 
 
-        public override void BackgroundInitialize() {
-            LoadingStatus = "Wasting CPU Cycles...";
+public class LoadingScreenTest : TestScreen {
+    public override bool RequireLoadingScreen => true;
 
-            Stopwatch stopwatch = new();
-            stopwatch.Start();
+    public override void BackgroundInitialize() {
+        LoadingStatus = "Wasting CPU Cycles...";
 
-            const float miliTotal = 5000;
+        Stopwatch stopwatch = new();
+        stopwatch.Start();
+
+        const float miliTotal = 5000;
             
-            while (stopwatch.Elapsed.TotalMilliseconds < miliTotal) {
-                LoadingProgress = (float)(stopwatch.Elapsed.TotalMilliseconds / miliTotal);
-                Thread.Sleep(1);
-            }
+        while (stopwatch.Elapsed.TotalMilliseconds < miliTotal) {
+            LoadingProgress = (float)(stopwatch.Elapsed.TotalMilliseconds / miliTotal);
+            Thread.Sleep(1);
+        }
             
-            LoadingComplete = true;
-            stopwatch.Stop();
-        }
+        LoadingComplete = true;
+        stopwatch.Stop();
+    }
 
-        public override void Initialize() {
-            base.Initialize();
+    public override void Initialize() {
+        base.Initialize();
 
-            this.Manager.Add(new TextDrawable(new Vector2(10), FurballGame.DEFAULT_FONT_STROKED, "Loading Complete!", 30));
-        }
+        this.Manager.Add(new TextDrawable(new Vector2(10), FurballGame.DEFAULT_FONT_STROKED, "Loading Complete!", 30));
     }
 }
