@@ -8,41 +8,41 @@ using Furball.Game.Screens;
 using Silk.NET.Input;
 using Color=Furball.Vixie.Backends.Shared.Color;
 
-namespace Furball.Game {
-    public class TestScreen : Screen {
-        private TexturedDrawable _background;
+namespace Furball.Game; 
+
+public class TestScreen : Screen {
+    private TexturedDrawable _background;
         
-        public override void Initialize() {
-            base.Initialize();
+    public override void Initialize() {
+        base.Initialize();
 
-            this.Manager.Add(this._background = new TexturedDrawable(FurballGame.WhitePixel, Vector2.Zero) {
-                ColorOverride = Color.BlueViolet,
-                Depth = 2
-            });
+        this.Manager.Add(this._background = new TexturedDrawable(FurballGame.WhitePixel, Vector2.Zero) {
+            ColorOverride = Color.BlueViolet,
+            Depth         = 2
+        });
 
-            this.Manager.Add(
-            new DrawableButton(
-            new(10, FurballGame.DEFAULT_WINDOW_HEIGHT - 10),
-            FurballGame.DEFAULT_FONT,
-            24,
-            "Back",
-            Color.Red,
-            Color.White,
-            Color.Black,
-            Vector2.Zero,
-            delegate {
-                ScreenManager.ChangeScreen(new ScreenSelector());
-            }
-            ) {
-                OriginType = OriginType.BottomLeft
-            }
-            );
+        this.Manager.Add(
+        new DrawableButton(
+        new(10, FurballGame.DEFAULT_WINDOW_HEIGHT - 10),
+        FurballGame.DEFAULT_FONT,
+        24,
+        "Back",
+        Color.Red,
+        Color.White,
+        Color.Black,
+        Vector2.Zero,
+        delegate {
+            ScreenManager.ChangeScreen(new ScreenSelector());
         }
-
-        public override void Relayout(float newWidth, float newHeight) {
-            base.Relayout(newWidth, newHeight);
-
-            this._background.Scale = new Vector2(newWidth, newHeight);
+        ) {
+            OriginType = OriginType.BottomLeft
         }
+        );
+    }
+
+    public override void Relayout(float newWidth, float newHeight) {
+        base.Relayout(newWidth, newHeight);
+
+        this._background.Scale = new Vector2(newWidth, newHeight);
     }
 }
